# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: IAM Service API
  version: 0.1.0
  description: |
    Аутентификация, организации, членство и RBAC. Префикс `/v1` может навешиваться API Gateway.
servers:
  - url: https://api.example.com/v1/iam
    description: через gateway
  - url: http://localhost:8080/v1/iam
    description: локально
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [auth]
      security: []
      summary: Регистрация по email и паролю
      operationId: authRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Вход по email и паролю
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Пара токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [auth]
      security: []
      summary: Обновление access по refresh (ротация refresh)
      operationId: authRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      security: []
      summary: Отзыв текущей refresh-сессии
      operationId: authLogout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Сессия отозвана
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/forgot-password:
    post:
      tags: [auth]
      security: []
      summary: Запрос сброса пароля (событие в notification)
      operationId: authForgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: Запрос принят (без раскрытия наличия email)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedEmpty'
  /auth/reset-password:
    post:
      tags: [auth]
      security: []
      summary: Сброс пароля по одноразовому токену
      operationId: authResetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Пароль обновлён
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/verify-email:
    post:
      tags: [auth]
      security: []
      summary: Подтверждение email
      operationId: authVerifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '204':
          description: Email подтверждён
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/switch-org:
    post:
      tags: [auth]
      summary: Выпуск новой пары токенов в контексте организации
      operationId: authSwitchOrg
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwitchOrgRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /auth/oidc/{provider}/start:
    get:
      tags: [auth]
      security: []
      summary: Редирект на внешний OIDC (Google, GitHub, …)
      operationId: authOidcStart
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            examples: [google, github]
      responses:
        '302':
          description: Redirect
  /auth/oidc/{provider}/callback:
    get:
      tags: [auth]
      security: []
      summary: Callback OIDC
      operationId: authOidcCallback
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        '302':
          description: Redirect в приложение с токенами или one-time code
  /auth/sso/{orgSlug}/start:
    get:
      tags: [auth]
      security: []
      summary: Старт Enterprise SSO для организации
      operationId: authSsoStart
      parameters:
        - name: orgSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect на IdP
  /auth/sso/{orgSlug}/acs:
    post:
      tags: [auth]
      security: []
      summary: SAML ACS (POST binding)
      operationId: authSsoSamlAcs
      parameters:
        - name: orgSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: true
      responses:
        '302':
          description: Redirect в приложение
  /me:
    get:
      tags: [me]
      summary: Текущий пользователь и организации
      operationId: meGet
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
    patch:
      tags: [me]
      summary: Обновление профиля
      operationId: mePatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MePatchRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
  /me/password:
    post:
      tags: [me]
      summary: Смена пароля
      operationId: meChangePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /me/sessions/{sessionId}:
    delete:
      tags: [me]
      summary: Отзыв сессии
      operationId: meRevokeSession
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: OK
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Forbidden:
      description: Нет прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Conflict:
      description: Конфликт (например, email занят)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    AcceptedEmpty:
      type: object
      properties:
        accepted:
          type: boolean
          const: true
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 10
        displayName:
          type: string
          maxLength: 200
        organizationName:
          type: string
          description: Если задано — создаётся org и пользователь становится owner
          maxLength: 200
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 10
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    SwitchOrgRequest:
      type: object
      required: [orgId]
      properties:
        orgId:
          type: string
          format: uuid
    AuthTokensResponse:
      type: object
      required: [accessToken, accessExpiresIn, tokenType]
      properties:
        accessToken:
          type: string
        accessExpiresIn:
          type: integer
          description: секунды до истечения access
        refreshToken:
          type: string
          description: Может отсутствовать при выдаче только cookie
        refreshExpiresIn:
          type: integer
        tokenType:
          type: string
          const: Bearer
    UserPublic:
      type: object
      required: [id, email, status]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        displayName:
          type: string
        locale:
          type: string
        status:
          type: string
          enum: [active, suspended, deleted]
    MeOrgSummary:
      type: object
      required: [orgId, slug, name, roles]
      properties:
        orgId:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        roles:
          type: array
          items:
            type: string
    MeResponse:
      type: object
      required: [user, organizations]
      properties:
        user:
          $ref: '#/components/schemas/UserPublic'
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/MeOrgSummary'
    MePatchRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 200
        locale:
          type: string
          maxLength: 16
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 10
...